refactor: clarify runtime config delivery semantics
